[RemainingToText] Move module level format functions to a helper file
